requirements.txt
selenium==4.15.2
tqdm==4.66.1
requests==2.31.0
beautifulsoup4==4.12.2
urllib3==2.0.7
webdriver-manager==4.0.1
fake-useragent==1.4.0
python-dotenv==1.0.0 